Variant consequence

This function maps consequences (CSQ) of a given list of variants to those in an annotated set based on facilities such as variant effect predictor (VEP). In the case of protein-altering variants (PAVs), many consequences can be involved. The procedure also takes into account linkage disequilibrium (LD) in flanking regions.

Usage

csq(
  query_loci,
  annotated_loci,
  pattern,
  ldops = NULL,
  flanking = 1e+06,
  pop = "EUR",
  verbose = TRUE
)

Arguments

query_loci

A data.frame of loci whose consequences are to be obtained.

annotated_loci

A data.frame of annotated loci.

ldops

arguments for ieugwasr::ld_matrix_local()

flanking

A flanking distance.

pop

The reference population as for ieugwasr::ld_matrix().

verbose

A flag to show nonexistent variants.

consequences

A set of consequences to check against.

Value

A data.frame with an indicator with/without those consequences.