QTL lookup

This function takes MR results (involving pQTL/QTL) to look up QTLs in trait GWASs given a P value cutoff. The rationale is that a pQTL may not necessarily be in strong LD with QTL but some other independent signal in the same region.

Usage

qtl_lookup(
  dat,
  panel = "1000Genomes",
  pthreshold = 0.001,
  pop = "EUR",
  plink_bin = NULL,
  r = NULL,
  r2 = NULL,
  xlsx = NULL
)

Arguments

dat

MR results (`protein`, `id`, `pqtl`,`p`, `qtl`, `p_qtl`) whose `proxy`, `p_proxy` and `rsq` variables will be updated.

panel

reference panel.

pthreshold

cutoff of QTL association from trait GWASs.

pop

pupulation.

plink_bin

PLINK executable file whose binary files are indicated in `bfile` variable of `dat`.

r

when specified, the LD(r) is output.

r2

when specified, the LD(r^2) is output.

xlsx

a non-null specification indicates name of an output Excel workbook.

Value

A data.frame containing the looked up loci.

Examples

if (FALSE) {
options(width=200)
INF <- Sys.getenv("INF")
suppressMessages(library(dplyr))
d <- file.path(INF,"mr","gsmr","trait")
gsmr_efo <- read.delim(file.path(INF,"mr","gsmr","gsmr-efo.txt")) %>%
            filter(fdr<=0.05) %>%
            mutate(prot=gsub("-",".",protein)) %>%
            mutate(file_gwas=paste(prot,id,"rsid.txt",sep="-"),
                   bfile=file.path(INF,"INTERVAL","per_chr",paste0("interval.imputed.olink.chr_",chr)),
                   proxy=NA,p_proxy=NA,rsq=NA)
proxies <- qtl_lookup(gsmr_efo,plink_bin="/rds/user/jhz22/hpc-work/bin/plink",xlsx=file.path(INF,"mr","gsmr","r2_INTERVAL.xlsx")) %>%
           select(protein,id,Disease,pqtl,p,qtl,p_qtl,proxy,p_proxy,rsq)
write.table(proxies,file=file.path(INF,"mr","gsmr","r2_INTERVAL.tsv"),row.names=FALSE,quote=FALSE,sep="\t")
}