phenoscanner regionqueries in batches R/phenoscanner only allows for certain number of items supplied. This simple function return a large number of calls in batches as well as generating SNPIDs.

phenoscanner regionqueries in batches R/phenoscanner only allows for certain number of items supplied. This simple function return a large number of calls in batches as well as generating SNPIDs.

regionqueries(
  regionlist,
  catalogue = "pQTL",
  proxies = "EUR",
  p = 5e-08,
  r2 = 0.8,
  build = 37,
  wait = TRUE
)

Arguments

regionlist

a list of SNPs

catalogue

"None","eQTL","mQTL","methQTL","pQTL","GWAS".

proxies

"None", "AFR","AMR","EAS","EUR","SAS".

p

p value threshold.

r2

r2 for LD.

build

37, 38.

wait

a flag to wait for 1hr for every 50 regions.

Value

The returned value is a list containing tiles, regions and results.

Details

Batches are generated and queries are combined into one.

Note

adapted from custom codings

References

Sun BB, et al. (2018). Genomic atlas of the human plasma proteome. Nature 558: 73-79.

See also

Examples