A generic wrapper for TwoSampleMR analysis

Given a compressed or uncompressed file containing exposure data, this function conducts a two-sample MR analysis using MR-Base.

run_TwoSampleMR(TwoSampleMRinput, plot=TRUE, prefix)

Arguments

TwoSampleMRinput
A list containing the following elements to allow for greater flexibility,

exposure result TwoSampleMR's format_data()

outcome result if TwoSampleMR's extract_outcome_data()

clump result of TwoSampleMR's clump_data()

harmonise result of TwoSampleMR's harmonise_data()

plot
to produce plots

prefix
a prefix for output files

Details

As TwoSampleMR includes seemingly daunting tasks, this function intends to simplify various steps in a two-sample MR. It is particularly useful when a large numbher of MRs are necessary, e.g., multiple proteins and their cis/trans regions need to be examined, in which case prefix could direct the output to various directories.

Value

No value is returned but several files.

References

Dimou NL, Tsilidis KK. A Primer in Mendelian Randomization Methodology with a Focus on Utilizing Published Summary Association Data. In Evangelos Evangelou (ed.), Genetic Epidemiology: Methods and Protocols, Methods in Molecular Biology, vol. 1793, https://doi.org/10.1007/978-1-4939-7868-7_13, Springer Science+Business Media, LLC, part of Springer Nature 2018

Note

Adapted from script by Dimou NL, Tsilidis KK

Examples

library(TwoSampleMR)
outcomes <- "ebi-a-GCST007432"
prot <- "MMP.10"
type <- "cis"
f <- paste0(prot,"-",type,".mrx")
d <- read.table(file.path(path.package("pQTLtools"),"tests",f),header=TRUE)

#> Warning: cannot open file '/home/jhz22/pQTLtools/tests/MMP.10-cis.mrx': No such file or directory

#> Error in file(file, "rt"): cannot open the connection

exposure <- format_data(within(d,{P=10^logP}), phenotype_col="prot", snp_col="rsid",
                        chr_col="Chromosome", pos_col="Posistion",
                        effect_allele_col="Allele1", other_allele_col="Allele2",
                        eaf_col="Freq1", beta_col="Effect", se_col="StdErr",
                        pval_col="P", log_pval=FALSE,
                        samplesize_col="N")

#> Error in within(d, {    P = 10^logP}): object 'd' not found

clump <- clump_data(exposure)

#> Error in is.data.frame(dat): object 'exposure' not found

outcome <- extract_outcome_data(snps=exposure$SNP,outcomes=outcomes)

#> Error in unique(snps): object 'exposure' not found

harmonise <- harmonise_data(clump,outcome)

#> Error in required_columns %in% names(dat): object 'clump' not found

input <- list(exposure,outcome,clump,harmonise)

#> Error in eval(expr, envir, enclos): object 'exposure' not found

run_TwoSampleMR(input, prefix=paste(outcomes,prot,type,sep="-"))

#> Error in with(TwoSampleMRinput, exposure): object 'input' not found