Manhattan plots

Manhattan plots

Usage

turboman(
  input_data_path,
  output_data_rootname,
  custom_peak_annotation_file_path,
  reference_file_path,
  pvalue_sign,
  plot_title
)

Arguments

input_data_path

Path of the input association data

output_data_rootname

Root name of the plot output file

custom_peak_annotation_file_path

Path of the custom annotation of variants

reference_file_path

Path to the "turboman_hg19_reference_data.rda" reference file

pvalue_sign

Significance threshold p-value

plot_title

Plot title which will be displayed on top of the plot

Details

Input association data path / input_data_path

Define the path of the input association data... The input data needs to be a file that has:

Spaces as field separators

One header line

Two option:

I (no extreme p-values present): 3 columns, being chromosome, position, pvalue in order, column names are not important

II (extreme p-values present): 5 columns, being chromosome, position, pvalue, beta, se in order, column names are not important

Output data root name / output_data_rootname

Define the root name of the plot output file... Ex.: "my_man_plot" will result in an output file named "my_man_plot.png"

OPTIONAL: Custom annotation file / custom_peak_annotation_file_path

Define the path of the custom annotation of variants... The input data needs to be a file that has:

Spaces / tabs as field separators

One header line with exact column names (order not important)

3 columns: chromosome, position, label (e.g., gene name) Column names chromosome, position, nearest_gene_name NB!: If no label is given, variants will be automatically annotated

Reference file path / reference_file_path

Define the path to the "turboman_hg19_reference_data.rda" reference file that contains the LD block breaks as in Berisa and Pickrell (2016) and gene coordinates used to construct and annotate the Manhattan plot

Significance threshold p-value / pvalue_sign

Define the significance threshold... This will be used to...

Highlight signal peaks that come above this significance threshold

Annotate the nearest gene to the top signal in the peak

Draw a horizontal reference line equal to this threshold

References

Berisa T, Pickrell JK (2016). “Approximately independent linkage disequilibrium blocks in human populations.” Bioinformatics, 32(2), 283-285. doi:10.1093/bioinformatics/btv546 .

Author

Arthur Gilly, Chris Finan, Bram Prins

Changelog

2017/12/04. Added Varying window size in which gene names are plotted based on the length of gene names so that gene names will not be truncated in the plot

2017/12/04. Added an option allowing the user to supply a table with top SNPs to highlight, which can, for example, be selected by conditional analyses / LD pruning

2017/12/04. Fixed a bug (found by Tao) that leads to erroneous binning when plotting extremely sparse data

2017/12/05. Added optional log-pval calculations from beta/SE for extreme p-values that run below .Machine$double.xmin

2017/12/05. Added an automated option for plotting already log-transformed p-values (user can supply either)

2017/12/05. Increased contrast between chromosomes on the plot (Eric H)

2017/12/05. Set '10' in the y-axis label (-log10 pvalue) to subscript (Eric H)

To do list

Display gene names bold / non-bold when genic / intergenic

Create LD blocks for hg38

Add many more sanity checks (are you giving pvalues, the file does not exist, etc.)

Create a function version of the script (Eric H)

...

Examples

if (FALSE) {
require(gap.datasets)
test <- mhtdata[c("chr","pos","p")]
write.table(test,file="test.txt",row.names=FALSE,quote=FALSE)
annotate <- unique(subset(mhtdata[c("chr","start","gene","p")],p<5e-8 & gene!=""))
names(annotate) <- c("chromosome","position","nearest_gene_name","p")
write.table(annotate[,-4],file="annotate.txt",row.names=FALSE,quote=FALSE)
input_data_path <- "test.txt"
output_data_rootname <- "test_man"
custom_peak_annotation_file_path <- "annotate.txt"
reference_file_path <- "turboman_hg19_reference_data.rda"
pvalue_sign <- 5e-8
plot_title <- "gap.datasets example"
turboman(input_data_path, output_data_rootname, custom_peak_annotation_file_path, reference_file_path, pvalue_sign, plot_title)

input_data_path <- "FCN3.txt.gz"
output_data_rootname <- "FCN3"
custom_peak_annotation_file_path <- "FCN3.txt"
reference_file_path <- "turboman_hg19_reference_data.rda"
pvalue_sign <- 5e-8
plot_title <- "FCN3"
turboman(input_data_path, output_data_rootname, custom_peak_annotation_file_path, reference_file_path, pvalue_sign, plot_title)
}