QQ plots

QQ plots

Usage

turboqq(input_data_path, output_data_rootname, plot_title)

Arguments

input_data_path

Path of the input association data.

output_data_rootname

Root name of the plot output file.

plot_title

Plot title to be displayed on top of the plot.

Value

No direct return value. The script generates QQ plots as output.

Details

The method uses the kth order statistic from a sample of n i.i.d. U(0,1) statistics has a Beta(k,n+1-k) distribution as in Quesenberry and Hales (1980) and Coded by Weale M, Price T. https://sites.google.com/site/mikeweale/software

Input association data path / input_data_path:

Define path of the input association data. The input data needs to be a file that has:

Spaces as field separators.

One header line.

Option I. (no extreme p-values present): 3 columns, being chromosome, position, pvalue in order, column names are not important. Option II. (extreme p-values present): 5 columns, being chromosome, position, pvalue, beta, se in order, column names are not important.

References

Quesenberry CP, Hales C (1980). “Concentration bands for uniformity plots.” J Statist Comput Simul, 11(1), 41-53. doi:10.1080/00949658008810388 .

Author

Bram Prins

Examples

if (FALSE) {
require(gap.datasets)
test <- mhtdata[c("chr","pos","p")]
write.table(test,file="test.txt",row.names=FALSE,quote=FALSE)
input_data_path <- "test.txt"
output_data_rootname <- "test_qq"
plot_title <- "gap.datasets example"
turboqq(input_data_path, output_data_rootname, plot_title)
}