A generic wrapper for TwoSampleMR analysis

Given a compressed or uncompressed file containing exposure data, this function conducts a two-sample MR analysis using MR-Base.

run_TwoSampleMR(exposure.args=format_file.args(),outcome.args=extract_outcome_data.args(),
                       clump.args=clump_data.args(),harmonise.args=harmonise_data.args(),prefix,...)

Arguments

exposure.args
a list of options appropriate for TwoSampleMR's format_data() in a file,

file. NULL (file)

type. exposure (or outcome)

snps. NULL

header. TRUE (or FALSE)

phenotype_col. "Phenotype" (protein column)

snp_col. "SNP"

beta_col. "beta"

se_col. "se"

eaf_col. "eaf"

effect_allele_col. "effect_allele"

other_allele_col. "other_allele"

pval_col. "pval"

units_col. "units"

ncase_col. "ncase"

ncontrol_col. "ncontrol"

samplesize_col. "samplesize"

gene_col. "gene"

id_col. "id"

min_pval. 1e-200

z_col. "z"

info_col. "info"

chr_col. "chr"

pos_col. "pos"

log_pval. FALSE (or TRUE)

outcome.args
a list of options appropriate for TwoSampleMR's extract_outcome_data(),

snp. NULL (to derive from exposure.args)

outcomes. NULL

proxies. TRUE or FALSE

rsq. r^2

align_alleles. 1

palindromes. 1

maf_threshold. 0.3

access_token. token

splitsize. 10000

proxy_splitsize. 500

clump.args
a list of options appropriate for TwoSampleMR's clump_data(),

dat. NULL (format_data() output)

clump_kb. 10000

clump_r2. 0.001

clump_p1. 1

clump_p2. 1

pop. "EUR" ("EUR", "SAS", "EAS", "AFR", "AMR")

harmonise.args
a list of options appropriate for harmonise_data(),

exposure_data. NULL (exposure data from format_data())

outcome_data. NULL (outcome data from format_data())

action. 2 (1, 2, 3)

prefix
a prefix for output files

...
other options to be added

Details

As TwoSampleMR includes seemingly daunting tasks, this function intends to simplify various steps in a two-sample MR. It is particularly useful when a large numbher of MRs are necessary, e.g., multiple proteins and their cis/trans regions need to be examined, in which case prefix could direct the output to various directories.

Value

No value is returned but several files.

References

Dimou NL, Tsilidis KK. A Primer in Mendelian Randomization Methodology with a Focus on Utilizing Published Summary Association Data. In Evangelos Evangelou (ed.), Genetic Epidemiology: Methods and Protocols, Methods in Molecular Biology, vol. 1793, https://doi.org/10.1007/978-1-4939-7868-7_13, Springer Science+Business Media, LLC, part of Springer Nature 2018

Note

Adapted from script by Dimou NL, Tsilidis KK

Examples

# SCALLOP-INF
INF <- Sys.getenv("INF")
outcomes <- "ukb-b-19732"
prot <- "MMP.10"
type <- "cis"
wd <- file.path(INF,"mr",type)
f <- paste0(prot,"-",type,".mrx")
d <- file.path(wd,f)
library(TwoSampleMR)

#> TwoSampleMR version 0.5.5 
#> [>] New: Option to use non-European LD reference panels for clumping etc
#> [>] Some studies temporarily quarantined to verify effect allele
#> [>] See news(package='TwoSampleMR') and https://gwas.mrcieu.ac.uk for further details

if (FALSE) {
run_TwoSampleMR(exposure.args=foramt_file.args(d),outcome.args=extract_outcome_data.args(outcomes=outcomes),
                clump.args=clump_data.args(),harmonise.args=harmonise_data.args(),
                prefix=file.path(wd,paste(outcomes,prot,type,sep="-")))
}
# pQTLtools
d <- file.path(find.package("pQTLtools"),"tests",f)
run_TwoSampleMR(exposure.args=format_file.args(d),outcome.args=extract_outcome_data.args(outcomes=outcomes),
                clump.args=clump_data.args(),harmonise.args=harmonise_data.args(),
                prefix=paste(outcomes,prot,type,sep="-"))

#> Error in format_file.args(d): object 'dat' not found